Update README.md
I changed the text to answer the questions in a similar format. This includes writing about me and my research interests and demonstrating my basic knowledge of the course material by adding a python code block, an in line python expression, an image, a link to the WVU Geography site, and multiple lists.